move tests from lot module to lot_blanks
