+++ openprocurement/tender/limited/tests/lot_blanks.py
@@ -1,0 +1,903 @@
+# -*- coding: utf-8 -*-
+from openprocurement.tender.belowthreshold.tests.base import test_organization
+
+# TenderLotNegotiationResourceTest
+
+
+def create_tender_lot_invalid(self):
+    response = self.app.post_json('/tenders/some_id/lots',
+                                  {'data': {'title': 'lot title', 'description': 'lot description'}},
+                                  status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location': u'url', u'name': u'tender_id'}
+    ])
+
+    request_path = '/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token)
+
+    response = self.app.post(request_path, 'data', status=415)
+    self.assertEqual(response.status, '415 Unsupported Media Type')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description':
+             u"Content-Type header should be one of ['application/json']",
+         u'location': u'header',
+         u'name': u'Content-Type'}
+    ])
+
+    response = self.app.post(request_path, 'data', content_type='application/json', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'No JSON object could be decoded',
+         u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, 'data', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+         u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, {'not_data': {}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+         u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'value'},
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'title'},
+    ])
+
+    response = self.app.post_json(request_path, {'data': {'invalid_field': 'invalid_value'}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Rogue field', u'location':
+            u'body', u'name': u'invalid_field'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {'value': 'invalid_value'}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [
+            u'Please use a mapping for this field or Value instance instead of unicode.'],
+            u'location': u'body',
+            u'name': u'value'}
+    ])
+
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {"items": [{'relatedLot': '0' * 32}]}},
+                                   status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'relatedLot': [u'relatedLot should be one of lots']}],
+         u'location': u'body',
+         u'name': u'items'}
+    ])
+
+
+def create_tender_lot(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertEqual(lot['title'], 'lot title')
+    self.assertEqual(lot['description'], 'lot description')
+    self.assertIn('id', lot)
+    self.assertIn(lot['id'], response.headers['Location'])
+    self.assertNotIn('guarantee', lot)
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': lot}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [u'Lot id should be uniq for all lots'], u'location': u'body', u'name': u'lots'}
+    ])
+
+
+def create_complete_tender_lot(self):
+    """ Can't create lot when tender status is complete """
+    self.set_status('complete')
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]},
+                                  status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u"location": u"body",
+         u"name": u"data",
+         u"description": u"Can't add lot in current (complete) tender status"}
+    ])
+
+
+def create_cancelled_tender_lot(self):
+    """ Can't create lot when tender status is cancelled """
+    self.set_status('cancelled')
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]},
+                                  status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u"location": u"body",
+         u"name": u"data",
+         u"description": u"Can't add lot in current (cancelled) tender status"}
+    ])
+
+
+def create_unsuccessful_tender_lot(self):
+    """ Can't create lot when tender status is unsuccessful """
+    self.set_status('unsuccessful')
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]},
+                                  status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u"location": u"body",
+         u"name": u"data",
+         u"description": u"Can't add lot in current (unsuccessful) tender status"}
+    ])
+
+
+def patch_tender_lot(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        {"data": {"title": "new title"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["title"], "new title")
+
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        {"data": {"description": "new description"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["description"], "new description")
+
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        {"data": {"value": {"amount": 400}}})
+
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["value"]["amount"], 400)
+
+    response = self.app.patch_json('/tenders/{}/lots/some_id'.format(self.tender_id),
+                                   {"data": {"title": "other title"}},
+                                   status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'lot_id'}
+    ])
+
+    response = self.app.patch_json('/tenders/some_id/lots/some_id', {"data": {"title": "other title"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'tender_id'}
+    ])
+
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["title"], "new title")
+
+    self.set_status('complete')
+
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        {"data": {"title": "other title"}},
+        status=403)
+
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"],
+                     "Can't update lot in current (complete) tender status")
+
+
+def patch_tender_currency(self):
+    # create lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertEqual(lot['value']['currency'], "UAH")
+
+    # update tender currency
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {"value": {"currency": "GBP"}}})
+
+    self.assertEqual(response.status, '200 OK')
+    # log currency is updated too
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertEqual(lot['value']['currency'], "GBP")
+
+    # try to update lot currency
+    response = self.app.patch_json(
+        '/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'], self.tender_token),
+        {"data": {"value": {"currency": "USD"}}})
+    self.assertEqual(response.status, '200 OK')
+    # but the value stays unchanged
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertEqual(lot['value']['currency'], "GBP")
+
+    # try to update lot minimalStep currency and lot value currency in single request
+    response = self.app.patch_json(
+        '/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'], self.tender_token),
+        {"data": {"value": {"currency": "USD"}}})
+    self.assertEqual(response.status, '200 OK')
+    # but the value stays unchanged
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertEqual(lot['value']['currency'], "GBP")
+
+
+def patch_tender_vat(self):
+    # set tender VAT
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {"value": {"valueAddedTaxIncluded": True}}})
+
+    self.assertEqual(response.status, '200 OK')
+
+    # create lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertTrue(lot['value']['valueAddedTaxIncluded'])
+
+    # update tender VAT
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {"value": {"valueAddedTaxIncluded": False}}})
+
+    self.assertEqual(response.status, '200 OK')
+    # log VAT is updated too
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertFalse(lot['value']['valueAddedTaxIncluded'])
+
+    # try to update lot VAT
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id,
+                                                                             lot['id'],
+                                                                             self.tender_token),
+                                   {"data": {"value": {"valueAddedTaxIncluded": True}}})
+    self.assertEqual(response.status, '200 OK')
+    # but the value stays unchanged
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertFalse(lot['value']['valueAddedTaxIncluded'])
+
+    # try to update value VAT in single request
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id,
+                                                                             lot['id'],
+                                                                             self.tender_token),
+                                   {"data": {"value": {"valueAddedTaxIncluded": True}}})
+    self.assertEqual(response.status, '200 OK')
+    # but the value stays unchanged
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, lot['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.assertFalse(lot['value']['valueAddedTaxIncluded'])
+
+
+def delete_unsuccessful_tender_lot(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'], lot)
+
+    response = self.app.delete('/tenders/{}/lots/some_id?acc_token={}'.format(
+        self.tender_id, self.tender_token),
+        status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'lot_id'}
+    ])
+
+    response = self.app.delete('/tenders/some_id/lots/some_id', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'tender_id'}
+    ])
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {
+                                       "items": [{'relatedLot': lot['id']}]}
+                                   })
+    self.assertEqual(response.status, '200 OK')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'relatedLot': [u'relatedLot should be one of lots']}],
+         u'location': u'body',
+         u'name': u'items'}
+    ])
+    self.set_status('unsuccessful')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"],
+                     "Can't delete lot in current (unsuccessful) tender status")
+
+    response = self.app.get('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'], self.tender_token))
+    self.assertEqual(response.status, '200 OK')
+
+
+def delete_tender_lot(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'], lot)
+
+    response = self.app.delete('/tenders/{}/lots/some_id?acc_token={}'.format(
+        self.tender_id, self.tender_token),
+        status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'lot_id'}
+    ])
+
+    response = self.app.delete('/tenders/some_id/lots/some_id', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'tender_id'}
+    ])
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {
+                                       "items": [{'relatedLot': lot['id']}]}})
+    self.assertEqual(response.status, '200 OK')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'relatedLot': [u'relatedLot should be one of lots']}],
+         u'location': u'body',
+         u'name': u'items'}
+    ])
+
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {
+                                       "items": [{'relatedLot': None}]}})
+    self.assertEqual(response.status, '200 OK')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token))
+
+    self.assertEqual(response.status, '200 OK')
+
+    response = self.app.get('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'], self.tender_token),
+                            status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'lot_id'}
+    ])
+
+
+def delete_complete_tender_lot(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'], lot)
+
+    response = self.app.delete('/tenders/{}/lots/some_id?acc_token={}'.format(
+        self.tender_id, self.tender_token),
+        status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'lot_id'}
+    ])
+
+    response = self.app.delete('/tenders/some_id/lots/some_id', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'tender_id'}
+    ])
+
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    response = self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                                   {"data": {
+                                       "items": [{'relatedLot': lot['id']}]}
+                                   })
+    self.assertEqual(response.status, '200 OK')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'relatedLot': [u'relatedLot should be one of lots']}],
+         u'location': u'body',
+         u'name': u'items'}
+    ])
+    self.set_status('complete')
+
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token),
+        status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"],
+                     "Can't delete lot in current (complete) tender status")
+
+    response = self.app.get('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'], self.tender_token))
+    self.assertEqual(response.status, '200 OK')
+
+
+def cancel_lot_after_sing_contract(self):
+    # Create lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    #  Update item with relatedLot field
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': lot['id']}]}})
+    # Create award
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(
+        self.tender_id, self.tender_token), {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                                      'qualified': True, 'value': {"amount": 469,
+                                                                                   "currency": "UAH",
+                                                                                   "valueAddedTaxIncluded": True},
+                                                      'lotID': lot['id']}})
+    self.assertEqual(response.status, '201 Created')
+    award = response.json['data']
+
+    # Activate award
+    response = self.app.patch_json('/tenders/{}/awards/{}?acc_token={}'.format(
+        self.tender_id, award['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    response = self.app.get('/tenders/{}/contracts'.format(self.tender_id))
+    contract = response.json['data'][0]
+
+    # time travel
+    tender = self.db.get(self.tender_id)
+    for i in tender.get('awards', []):
+        if i.get('complaintPeriod', {}):  # reporting procedure does not have complaintPeriod
+            i['complaintPeriod']['endDate'] = i['complaintPeriod']['startDate']
+    self.db.save(tender)
+
+    # Activate contract
+    response = self.app.patch_json('/tenders/{}/contracts/{}?acc_token={}'.format(
+        self.tender_id, contract['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # try to cancel lot
+    response = self.app.post_json('/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {"data": {"reason": "cancellation reason", "cancellationOf": "lot",
+                                            "relatedLot": lot['id']}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.json['errors'], [
+        {u"location": u"body",
+         u"name": u"data",
+         u"description": u"Can't add cancellation in current (complete) tender status"}])
+
+
+def cancel_lot_with_complaint(self):
+    # Create lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': lot['id']}]}})
+
+    # Create award
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(
+        self.tender_id, self.tender_token), {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                                      'qualified': True, 'value': {"amount": 469,
+                                                                                   "currency": "UAH",
+                                                                                   "valueAddedTaxIncluded": True},
+                                                      'lotID': lot['id']}})
+    self.assertEqual(response.status, '201 Created')
+    award = response.json['data']
+    response = self.app.patch_json('/tenders/{}/awards/{}?acc_token={}'.format(
+        self.tender_id, award['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # Create complaints
+    response = self.app.post_json('/tenders/{}/awards/{}/complaints'.format(self.tender_id, award['id']),
+                                  {"data": {"status": "pending", "title": "Недостатньо інформації",
+                                            "description": "Опис",
+                                            "author": {"contactPoint": {"telephone": "+380 (432) 21-69-30",
+                                                                        "name": "Сергій Олексюк",
+                                                                        "email": "soleksuk@gmail.com"},
+                                                       "identifier": {"scheme": "UA-EDR",
+                                                                      "legalName": "Державне комунальне підприємство",
+                                                                      "id": "13313462",
+                                                                      "uri": "http://sch10.edu.vn.ua/"},
+                                                       "name": "ДКП «Школяр»",
+                                                       "address": {"countryName": "Україна", "postalCode": "21100",
+                                                                   "region": "м. Вінниця",
+                                                                   "streetAddress": "вул. Островського, 33",
+                                                                   "locality": "м. Вінниця"}}}})
+    self.assertEqual(response.status, '201 Created')
+
+    # Try to cancel lot
+    response = self.app.post_json('/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {"data": {"reason": "cancellation reason", "cancellationOf": "lot",
+                                            "relatedLot": lot['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.json['data']['status'], 'pending')
+    cancellation = response.json['data']
+    response = self.app.patch_json('/tenders/{}/cancellations/{}?acc_token={}'.format(
+        self.tender_id, cancellation['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # Check lot status
+    response = self.app.get('/tenders/{}/lots'.format(self.tender_id))
+    self.assertEqual(response.json['data'][0]['status'], 'cancelled')
+
+
+def last_lot_complete(self):
+    # Create 3 lots and update related items
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    first_lot = response.json['data']
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    second_lot = response.json['data']
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    third_lot = response.json['data']
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': first_lot['id']}]}})
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': second_lot['id']}]}})
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': third_lot['id']}]}})
+
+    # Cancel 1 lot
+    response = self.app.post_json('/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {"data": {"reason": "cancellation reason", "cancellationOf": "lot",
+                                            "relatedLot": first_lot['id']}})
+    self.app.patch_json('/tenders/{}/cancellations/{}?acc_token={}'.format(self.tender_id,
+                                                                           response.json['data']['id'],
+                                                                           self.tender_token),
+                        {"data": {"status": "active"}})
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, first_lot['id']))
+    self.assertEqual(response.json['data']['status'], 'cancelled')
+
+    # create awards for second and third
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(
+        self.tender_id, self.tender_token), {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                                      'qualified': True, 'value': {"amount": 469,
+                                                                                   "currency": "UAH",
+                                                                                   "valueAddedTaxIncluded": True},
+                                                      'lotID': second_lot['id']}})
+    response = self.app.patch_json('/tenders/{}/awards/{}?acc_token={}'.format(
+        self.tender_id, response.json['data']['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(
+        self.tender_id, self.tender_token), {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                                      'qualified': True, 'value': {"amount": 469,
+                                                                                   "currency": "UAH",
+                                                                                   "valueAddedTaxIncluded": True},
+                                                      'lotID': third_lot['id']}})
+    response = self.app.patch_json('/tenders/{}/awards/{}?acc_token={}'.format(
+        self.tender_id, response.json['data']['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # time travel
+    tender = self.db.get(self.tender_id)
+    for i in tender.get('awards', []):
+        if i.get('complaintPeriod', {}):  # reporting procedure does not have complaintPeriod
+            i['complaintPeriod']['endDate'] = i['complaintPeriod']['startDate']
+    self.db.save(tender)
+
+    # Sign in contracts
+    response = self.app.get('/tenders/{}/contracts'.format(self.tender_id))
+    first_contract = response.json['data'][0]
+    second_contract = response.json['data'][1]
+    response = self.app.patch_json('/tenders/{}/contracts/{}?acc_token={}'.format(
+        self.tender_id, first_contract['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+    response = self.app.patch_json('/tenders/{}/contracts/{}?acc_token={}'.format(
+        self.tender_id, second_contract['id'], self.tender_token), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # Check tender status
+    response = self.app.get('/tenders/{}'.format(self.tender_id))
+    self.assertEqual(response.json['data']['status'], 'complete')
+
+
+def all_cancelled_lots(self):
+    # Create 2 lots and update related items
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    first_lot = response.json['data']
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    second_lot = response.json['data']
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': first_lot['id']}]}})
+    self.app.patch_json('/tenders/{}?acc_token={}'.format(self.tender_id, self.tender_token),
+                        {'data': {'items': [{'relatedLot': second_lot['id']}]}})
+
+    # Cancel lots
+    response = self.app.post_json('/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {"data": {"reason": "cancellation reason", "cancellationOf": "lot",
+                                            "relatedLot": first_lot['id']}})
+    self.app.patch_json('/tenders/{}/cancellations/{}?acc_token={}'.format(self.tender_id,
+                                                                           response.json['data']['id'],
+                                                                           self.tender_token),
+                        {"data": {"status": "active"}})
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, first_lot['id']))
+    self.assertEqual(response.json['data']['status'], 'cancelled')
+    response = self.app.post_json('/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {"data": {"reason": "cancellation reason", "cancellationOf": "lot",
+                                            "relatedLot": second_lot['id']}})
+    self.app.patch_json('/tenders/{}/cancellations/{}?acc_token={}'.format(self.tender_id,
+                                                                           response.json['data']['id'],
+                                                                           self.tender_token),
+                        {"data": {"status": "active"}})
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, second_lot['id']))
+    self.assertEqual(response.json['data']['status'], 'cancelled')
+
+    # Check tender status
+    response = self.app.get('/tenders/{}'.format(self.tender_id))
+    self.assertEqual(response.json['data']['status'], 'cancelled')
+
+
+def cancel_lots_check_awards(self):
+    # create first lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    first_lot = response.json['data']
+
+    # create second lot
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    second_lot = response.json['data']
+
+    # first award
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                            'lotID': first_lot['id'], 'qualified': True}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    first_award = response.json['data']
+
+    # second award
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': {'suppliers': [test_organization], 'status': 'pending',
+                                            'lotID': second_lot['id'], 'qualified': True}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    second_award = response.json['data']
+
+    # cancellation for first lot
+    response = self.app.post_json(
+        '/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+        {"data": {"reason": "cancellation reason",
+                  "cancellationOf": "lot",
+                  "relatedLot": first_lot['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.json['data']['status'], 'pending')
+    cancellation = response.json['data']
+    response = self.app.patch_json('/tenders/{}/cancellations/{}?acc_token={}'.format(self.tender_id,
+                                                                                      cancellation['id'],
+                                                                                      self.tender_token),
+                                   {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # check corresponding lot
+    response = self.app.get('/tenders/{}/lots/{}'.format(self.tender_id, first_lot['id']))
+    self.assertEqual(response.json['data']['status'], 'cancelled')
+
+    # check corresponding award
+    response = self.app.get('/tenders/{}/awards/{}'.format(self.tender_id, first_award['id']))
+    self.assertEqual(response.json['data']['status'], 'pending')
+
+
+def delete_lot_after_first_award(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+    response = self.app.post_json('/tenders/{}/awards?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': {'suppliers': [test_organization], 'status': 'pending', 'lotID': lot['id'],
+                                            'qualified': True}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+    response = self.app.patch_json(
+        '/tenders/{}/awards/{}?acc_token={}'.format(self.tender_id, award['id'], self.tender_token),
+        {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']['status'], 'active')
+
+    # try to delete lot
+    response = self.app.delete('/tenders/{}/lots/{}?acc_token={}'.format(
+        self.tender_id, lot['id'], self.tender_token), status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [{
+        u"location": u"body",
+        u"name": u"data",
+        u"description": u"Can't delete lot when you have awards"}])
+
+
+def patch_lot_with_cancellation(self):
+    response = self.app.post_json('/tenders/{}/lots?acc_token={}'.format(self.tender_id, self.tender_token),
+                                  {'data': self.test_lots[0]})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    lot = response.json['data']
+
+    # Create cancellation on lot
+    response = self.app.post_json(
+        '/tenders/{}/cancellations?acc_token={}'.format(self.tender_id, self.tender_token),
+        {"data": {"reason": "cancellation reason",
+                  "cancellationOf": "lot",
+                  "relatedLot": lot['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.json['data']['status'], 'pending')
+
+    # Try to patch lot with cancellation on it
+    response = self.app.patch_json('/tenders/{}/lots/{}?acc_token={}'.format(self.tender_id, lot['id'],
+                                                                             self.tender_token),
+                                   {"data": {"title": "new title"}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.json['errors'][0]["description"],
+                     "Can't update lot when it has 'pending' cancellation.")
